Repository: Respond to boolean fields ending in question mark
If a field is a boolean field and is accessed like a boolean field, its
value will be fetched from the underlying data.

For example, the underlying data has a value for the "fork" key, which
is a boolean. This value can be retrieved using `repo.fork?` instead of
just `repo.fork`, without the need to explicitly define
`Repository#fork?`, as before. (In other words, this method of access
will now work with _all_ boolean fields.)