Mock upstream repo data
The upstream owner of dnsimple-python changed to tbunnyman (from
whatever it was...mike something), and that caused a whole bunch of
tests to fail. It was then that I realized I hadn't been properly
mocking HTTP calls when retrieving the upstream repo data.

Really, I should just create a mock _client_ that will handle this
transparently, but for now, this hacky way of doing it will work fine.